update smb proto
